+++ packages/overlay-api-client/src/projects/types.ts
@@ -4,4 +4,5 @@ export interface ProjectQuery extends PaginationQuery {
   projectId?: string
   updatedSince?: number
   includeDeleted?: boolean
+  archived?: boolean
 }